fix: revert the modification of student-scholarship schema
